#90 Create ShooterRampUp Button
Created a new constructor in ShooterRampUp that takes in a boolean. In
this constructor, the power is set up a value from preferences with key
SHOOTER_RAMP_UP_TEST_POWER. This key was then created in the preference
key class.
Next created a button on SmartDashboard that runs the ShooterRampUp
Command