Implement limit switch code #113
Change ManualRaiseAcquirer to use the rawRaiseArm method in Acquirer.

In Acquirer, add two methods: hasReachedUpperLimit and
hasReachedLowerLimit. Each one should return a boolean.
hasReachedUpperLimit should return if the upper limit switch (find in
RobotMap) is false. hasReachedLowerLimit should return if the lower
limit switch is false.

Change the rawRaiseArm method to set power equal to 0 if it
hasReachedUpperLimit and power is above 0. It should also set power
equal to 0 if it hasReachedLowerLimit and power is below 0. #113